fix: remove code used for migration of db after use
