+++ server/server.js
@@ -110,12 +110,11 @@ const { addPaths } = require('kth-node-express-routing')
 
 const { createApiPaths, createSwaggerRedirectHandler, notFoundHandler, errorHandler } = require('kth-node-api-common')
 const swaggerData = require('../swagger.json')
-const { System, MigrateMemo } = require('./controllers')
+const { System, StoredMemoPdf } = require('./controllers')
 
 // System pages routes
 const systemRoute = AppRouter()
-systemRoute.get('system.delete', config.proxyPrefixPath.uri + '/_delete', MigrateMemo.emptyCollection)
-systemRoute.get('system.count', config.proxyPrefixPath.uri + '/_count', MigrateMemo.collectionLength)
+systemRoute.get('system.count', config.proxyPrefixPath.uri + '/_count', StoredMemoPdf.collectionLength)
 systemRoute.get('system.monitor', config.proxyPrefixPath.uri + '/_monitor', System.monitor)
 systemRoute.get('system.about', config.proxyPrefixPath.uri + '/_about', System.about)
 systemRoute.get('system.paths', config.proxyPrefixPath.uri + '/_paths', System.paths)
@@ -147,7 +146,7 @@ addPaths(
 const authByApiKey = passport.authenticate('apikey', { session: false })
 
 // Application specific API enpoints
-const { Sample, CourseMemo, PDF, StoredMemoPdf } = require('./controllers')
+const { Sample, CourseMemo, PDF } = require('./controllers')
 const { ApiRouter } = require('kth-node-express-routing')
 
 const apiRoute = ApiRouter(authByApiKey)
@@ -158,8 +157,6 @@ apiRoute.register(paths.api.checkAPIkey, System.checkAPIKey)
 
 apiRoute.register(paths.api.getDataById, Sample.getData)
 apiRoute.register(paths.api.postDataById, Sample.postData)
-// Migration tool for data from kurs-pm-api
-apiRoute.register(paths.api.migrateAllData, MigrateMemo.migrateMemoInfoOfStoredPdf) // step 4: publish new version and unpublish prev version if it exists
 // Get one draft | update it
 apiRoute.register(paths.api.getDraftByEndPoint, CourseMemo.getDraftByEndPoint) // step 2: editor, fetch data
 apiRoute.register(paths.api.updateCreatedDraft, CourseMemo.putDraftByEndPoint) // step 2: editor, fast update
