+++ swagger.json
@@ -425,45 +425,10 @@
         ]
       }
     },
-    "/v1/migrateAll": {
-      "post": {
-        "operationId": "migrateAllData",
-        "summary": "Save migrated data from kurs-pm-api",
-        "description": "Save migrated data from kurs-pm-api",
-        "parameters": [
-          {
-            "name": "migratedDataArray",
-            "in": "body",
-            "description": "Array with memo data objects",
-            "required": true,
-            "schema": {
-              "$ref": "#/definitions/MigratedDataArray"
-            }
-          }
-        ],
-        "tags": ["v1"],
-        "responses": {
-          "200": {
-            "description": "A successful save"
-          },
-          "default": {
-            "description": "Unexpected error",
-            "schema": {
-              "$ref": "#/definitions/Error"
-            }
-          }
-        },
-        "security": [
-          {
-            "api_key": ["write"]
-          }
-        ]
-      }
-    },
     "/v1/storedCourseMemoFilesList/{courseCode}/{semester}": {
       "get": {
         "operationId": "getStoredMemoPdfListByCourseCode",
-        "summary": "Get a list of course memos stored as pdf files (integrated from kurs-pm-api)",
+        "summary": "Get a list of course memos stored as pdf files (integrated from kurs-pm-api) used by kursinfo-web",
         "description": "Gets course memos for rounds in a course. Memos stored as pdfs in kursinfo storage (not generated)",
         "parameters": [
           {
@@ -1038,64 +1003,6 @@
         "teacherAssistants": {
           "type": "string",
           "description": "Text field for html with Teacher assistants / Lärarassistenter (admin via kopps, fetched from redis-ug)"
-        }
-      }
-    },
-    "MigratedDataArray": {
-      "properties": {
-        "documents": {
-          "type": "array",
-          "items": {
-            "type": "object",
-            "description": "....",
-            "items": {
-              "schema": {
-                "$ref": "#/definitions/MigratedMemoInfoOfPdfSavedInStorage"
-              }
-            }
-          },
-          "description": "List of round ids that have been used for a course semester"
-        }
-      }
-    },
-    "MigratedMemoInfoOfPdfSavedInStorage": {
-      "properties": {
-        "_id": {
-          "type": "string",
-          "description": "Unique identifier representing some specific data"
-        },
-        "courseCode": {
-          "type": "string",
-          "description": "Course code for which it is valid"
-        },
-        "pdfMemoUploadDate": {
-          "type": "string",
-          "description": "When document were published (first)?? time"
-        },
-        "changedBy": {
-          "type": "string",
-          "description": "Unique KTH id of user who uploaded or changed this document data, f.e., u1iooii,"
-        },
-        "semester": {
-          "type": "string",
-          "description": "Course start term year (spring(VT)-1, autumn(HT)-2) , f.e. 20182 or 20181 (termin)"
-        },
-        "koppsRoundId": {
-          "type": "string",
-          "description": "List of round id:s for which memo is valid (kurstillfälle)"
-        },
-        "courseMemoFileName": {
-          "type": "string",
-          "description": "Uploaded courseMemo file"
-        },
-        "lastChangeDate": {
-          "type": "string",
-          "description": "Date and time for last change, set by API"
-        },
-        "previousFileList": {
-          "type": "array",
-          "items": "string",
-          "description": "Array for object with previous memo file and publish date "
         }
       }
     },
